Token Authentication - Seperated login API from rest of the APIs - Added token authentication middleware for CRUD APIs
