+++ Controllers/Admin.js
@@ -3,46 +3,6 @@ import jwt from "jsonwebtoken";
 import bcrypt from 'bcrypt';
 import { configDotenv } from "dotenv";
 
-configDotenv();
-const adminLogin = async (req, res) => {
-    const query = "SELECT password FROM admin WHERE email = ?";
-    try {
-
-        const [result, fields] = await con.query(query, [req.body.username, req.body.password]);
-        if (result.length > 0) {
-            const isPasswordCorrect = await bcrypt.compare(req.body.password, result[0].password);
-            
-            if (isPasswordCorrect) {
-                const user = result[0];
-                const email = user.email;
-                const token = jwt.sign({ role: "admin", email }, process.env.JWT_SECRET_KEY, { expiresIn: process.env.TOKEN_SPAN});
-                res.cookie('token', token);
-
-                return res.json({
-                    loginStatus: true,
-                });
-            } else {
-                res.status(401);
-                return res.json({
-                    error: 'Password did not match'
-                })
-            }
-        } else {
-            res.status(401);
-            return res.json({
-                loginStatus: false,
-                error: "User not found. Please check credentials.",
-            });
-        }
-    } catch (e) {
-        res.status(500);
-        return res.json({
-            loginStatus: false,
-            error: e,
-        });
-    }
-};
-
 const addCategory = async (req, res) => {
     const query = "INSERT INTO category (`name`) VALUE (?)";
     try {
@@ -90,6 +50,8 @@ const addEmployee = async (req, res) => {
 };
 
 const getAllEmployees = async (req, res) => {
+    // const authorizationHeader = req.headers.authorization;
+    // console.log('#### Calling Alll', authorizationHeader);
     try {
         const sql = 'SELECT name, email, address, salary, category_id FROM employees';
         const [results] = await con.query(sql);
@@ -201,7 +163,6 @@ const adminLogout = async (req, res) => {
 };
 
 export {
-    adminLogin,
     adminLogout,
     addCategory,
     addEmployee,
